Move installer tests to subfolder
